+++ lib/src/components/pin_pad.dart
@@ -2,20 +2,28 @@ import '../ui_bits_internal.dart';
 
 class BitPinPad extends StatelessWidget {
   final void Function(String) onTap;
+  final void Function() onTapBack;
+  final void Function() onTapClear;
   final Field<String> pinField;
+  final BitAnimation animation;
 
   const BitPinPad({
     Key key,
     this.onTap = _defaultTap,
     this.pinField,
+    this.animation = const BitNoAnimation(),
+    this.onTapBack,
+    this.onTapClear,
   }) : super(key: key);
 
   static void _defaultTap(String _) {}
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: _makeRows(context, this.pinField ?? Field.asText()),
+    return animation.wrapWidget(
+      child: Column(
+        children: _makeRows(context, this.pinField ?? Field.asText()),
+      ),
     );
   }
 
@@ -43,7 +51,10 @@ class BitPinPad extends StatelessWidget {
               children: [
                 BitCircleButton(
                   iconData: FontAwesomeIcons.chevronLeft,
-                  onTap: () => onTap.call('back'),
+                  onTap: () {
+                    onTap.call('back');
+                    onTapBack?.call();
+                  },
                 ),
                 SizedBox(width: context.sizes.medium),
                 BitCircleButton(
@@ -59,6 +70,7 @@ class BitPinPad extends StatelessWidget {
                   onTap: () {
                     field.setValue('');
                     onTap.call('clean');
+                    onTapBack?.call();
                   },
                 ),
               ],