+++ sample_app/lib/pin_pad_page_sample.dart
@@ -17,6 +17,7 @@ class PinPadPageSample extends StatelessWidget {
           ),
           SizedBox(height: context.sizes.extraLarge),
           BitPinPad(
+            animation: BitAnimations.scale(),
             pinField: pinField,
           ),
         ],