+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.20]
+- Allow animations on BitPinPad
+- Allow animations on BitCarousel
+- All animations receive onComplete
+  
 ## [0.0.19]
 - Field<T> should be accessed via factory method:
   - Field.asInt()